comfyui_custom_nodes/ComfyUI-Image-Scorer/requirements.txt
numpy
onnxruntime-gpu; sys_platform != "darwin"
onnxruntime; sys_platform == "darwin"
transformers
sentence-transformers
scikit-learn
Pillow
torch
